Run install checks before display, test installer status
